requirements.txt
streamlit==1.45.0
psycopg2-binary
pandas
xlsxwriter
openpyxl
authlib>=1.3.2
bcrypt==4.1.2
plotly>=5.19
Pillow>=10.2            # already present, just keep
pytest>=7.0
streamlit-autorefresh      # ← NEW  (no pinned version; pip will grab the latest)
